Changed the variable type to bigint from int

+++ CreateKobayashiIoT.sql
@@ -5,13 +5,13 @@ use kobayashi_iot;
 create user sopackuser@localhost identified by 'sopa-0001';
 grant all on kobayashi_iot.* to sopackuser@localhost;
 create table current (
-    id int unsigned not null primary key auto_increment,
+    id bigint unsigned not null primary key auto_increment,
     area int unsigned not null,
     datetime datetime not null,
     current_switch boolean not null default 0
 );
 create table environment (
-    id int unsigned not null primary key auto_increment,
+    id bigint unsigned not null primary key auto_increment,
     area int unsigned not null,
     datetime datetime not null,
     temperature int,
@@ -19,7 +19,7 @@ create table environment (
     illumination int
 );
 create table dust (
-    id int unsigned not null primary key auto_increment,
+    id bigint unsigned not null primary key auto_increment,
     area int unsigned not null,
     datetime datetime not null,
     dust int
